requirements.txt
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.25.2
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
ImageIO==2.37.3
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy-loader==0.5
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.17.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.3
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pillow==12.1.1
protobuf==6.33.5
pyarrow==23.0.1
pyclipper==1.4.0
pydeck==0.9.1
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.17.1
setuptools==82.0.1
shapely==2.1.2
six==1.17.0
smmap==5.0.3
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
tifffile==2026.3.3
toml==0.10.2
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
Unidecode==1.4.0
urllib3==2.6.3
watchdog==6.0.0
